Chng

+++ main.py
@@ -2,8 +2,8 @@ import concurrent.futures
 from pathlib import Path
 from sys import argv
 from threading import Thread
-from os import mkdir, rename
-from time import time, sleep
+from os import mkdir, replace, listdir
+from time import time
 
 timer = time()
 
@@ -37,7 +37,7 @@ def move_file(file: list):
     # thread = Thread(target=rename, args=(file, new_file_path))
     # thread.start()
     # threads.append(thread)
-    rename(file[0], new_file)
+    replace(file[0], new_file)
 
 
 def del_dirs(path):
@@ -45,12 +45,11 @@ def del_dirs(path):
         if directory.is_dir() and directory.name in folder_lst:
             continue
         else:
-            directory.rmdir()
-
-
-def del_dir(path, thread):
-    thread.join()
-    path.rmdir()
+            while True:
+                dir_list = list(directory.iterdir())
+                if len(dir_list) == 0:
+                    directory.rmdir()
+                    break
 
 
 def same_file_check(file):
@@ -70,14 +69,11 @@ def same_file_check(file):
 
 def handler(path):
     path_lst = []
-    threads = []
 
     for file in path.iterdir():
         if file.is_dir() and file.name not in folder_lst:
             thread = Thread(target=handler, args=(file, ))
             thread.start()
-            del_dir(file, thread)
-            # threads.append(thread)
             continue
         elif file.is_dir() and file.name in folder_lst:
             continue
@@ -85,12 +81,11 @@ def handler(path):
         cat = get_category(file)
         new_path = make_file_path(file, cat)
         path_lst.append([file, new_path])
-    executor = concurrent.futures.ThreadPoolExecutor(max_workers=1000)
+    executor = concurrent.futures.ThreadPoolExecutor(max_workers=50)
     futures = [executor.submit(move_file, path) for path in path_lst]
     done, not_done = concurrent.futures.wait(futures, return_when=concurrent.futures.ALL_COMPLETED)
 
-    # [el.join() for el in threads]
-    # del_dirs(path)
+    del_dirs(path)
         
 
 if __name__ == "__main__":
@@ -100,7 +95,5 @@ if __name__ == "__main__":
         print("Use path as an argument")
         exit()
     main_path = Path(" ".join(argv[1:]))
-    main_thread = Thread(target=handler, args=(main_path, ))
-    main_thread.start()
-    main_thread.join()
+    handler(main_path)
     print(time() - timer)
